refactor(furuno): consolidate protocol constants into protocol.rs
Collect all wire-level constants, command IDs, wire index tables,
frame header bitmasks, and model identification into a dedicated
protocol.rs module, following the pattern established by Navico.

Inline magic numbers in the frame header parser and spoke decoders
are replaced with named constants for readability. The parse_model()
function moves to RadarModel::from_part_number(). No behavioral
changes.
