+++ src/lib/brand/furuno/command.rs
@@ -1,201 +1,16 @@
 use async_trait::async_trait;
-use enum_primitive_derive::Primitive;
 use std::fmt::Write;
 use tokio::io::{AsyncWriteExt, WriteHalf};
 use tokio::net::TcpStream;
 
-use super::CommandMode;
+use super::protocol::{
+    CommandId, CommandMode, WIRE_UNIT_KM, WIRE_UNIT_NM, meters_to_wire_index_for_unit,
+    wire_unit_for_meters,
+};
 use crate::brand::CommandSender;
 use crate::radar::range::Ranges;
 use crate::radar::settings::{ControlId, ControlValue, SharedControls};
 use crate::radar::{Power, RadarError, RadarInfo};
-
-#[derive(Primitive, PartialEq, Eq, Debug, Clone)]
-pub(crate) enum CommandId {
-    Connect = 0x60,
-    DispMode = 0x61,
-    Range = 0x62,
-    Gain = 0x63,
-    Sea = 0x64,
-    Rain = 0x65,
-    CustomPictureAll = 0x66,
-    SignalProcessing = 0x67, // Multi-purpose: NoiseReduction, InterferenceRejection, etc.
-    PulseWidth = 0x68,
-    Status = 0x69,
-    U6D = 0x6D,
-    AntennaType = 0x6E,
-
-    Tune = 0x75,
-    TuneIndicator = 0x76,
-    BlindSector = 0x77,
-    DRS4WHeartbeat = 0x7D, // DRS4W-specific heartbeat (every ~1s)
-
-    Att = 0x80,
-    MainBangSize = 0x83,
-    AntennaHeight = 0x84,
-    NearSTC = 0x85,
-    MiddleSTC = 0x86,
-    FarSTC = 0x87,
-    RingSuppression = 0x88,
-    ScanSpeed = 0x89,
-    AntennaSwitch = 0x8A,
-    AntennaNo = 0x8D,
-    OnTime = 0x8E,
-    TxTime = 0x8F,
-
-    Modules = 0x96,
-
-    Drift = 0x9E,
-    TrailMode = 0xA3,
-    ConningPosition = 0xAA,
-    WakeUpCount = 0xAC,
-    Heartbeat = 0xAF, // $NAF,256 — frequent radar heartbeat
-
-    STCRange = 0xD2,
-    CustomMemory = 0xD3,
-    BuildUpTime = 0xD4,
-    DisplayUnitInformation = 0xD5,
-    TrailProcess = 0xE1,
-    CustomATFSettings = 0xE0,
-    AliveCheck = 0xE3,
-    ATFSettings = 0xEA,
-    BirdMode = 0xED,
-    RezBoost = 0xEE,       // Target Separation (beam sharpening)
-    TargetAnalyzer = 0xEF, // Doppler mode
-    AutoAcquire = 0xF0,
-    NN3Command = 0xF5, // Hardware diagnostics (frequent, read-only)
-    RangeSelect = 0xFE,
-}
-
-/// Furuno wire index to meters mapping table (NM mode, wire unit 0)
-/// CRITICAL: Wire indices are NON-SEQUENTIAL! The radar uses specific wire index values.
-/// Verified via Wireshark captures from TimeZero ↔ DRS4D-NXT
-///
-/// Example: To set 1/16nm range, you send wire_index=21 (NOT 0!)
-///          To set 36nm range, you send wire_index=19 (NOT 16!)
-pub const WIRE_INDEX_TABLE: [(i32, i32); 22] = [
-    (21, 116),    // 1/16 nm = 116m - wire index 21!
-    (0, 231),     // 1/8 nm = 231m
-    (1, 463),     // 1/4 nm = 463m
-    (2, 926),     // 1/2 nm = 926m
-    (3, 1389),    // 3/4 nm = 1389m
-    (4, 1852),    // 1 nm = 1852m
-    (5, 2778),    // 1.5 nm = 2778m
-    (6, 3704),    // 2 nm = 3704m
-    (7, 5556),    // 3 nm = 5556m
-    (8, 7408),    // 4 nm = 7408m
-    (9, 11112),   // 6 nm = 11112m
-    (10, 14816),  // 8 nm = 14816m
-    (11, 22224),  // 12 nm = 22224m
-    (12, 29632),  // 16 nm = 29632m
-    (13, 44448),  // 24 nm = 44448m
-    (14, 59264),  // 32 nm = 59264m
-    (19, 66672),  // 36 nm = 66672m (OUT OF SEQUENCE!)
-    (15, 88896),  // 48 nm = 88896m
-    (20, 118528), // 64 nm = 118528m (OUT OF SEQUENCE!)
-    (16, 133344), // 72 nm = 133344m
-    (17, 177792), // 96 nm = 177792m
-    (18, 222240), // 120 nm = 222240m
-];
-
-/// Furuno wire index to meters mapping table (km mode, wire unit 1)
-/// Same wire indices, but range values represent km instead of NM.
-/// Wire index 21 (0.0625) is NOT available in km mode.
-pub const WIRE_INDEX_TABLE_KM: [(i32, i32); 21] = [
-    (0, 125),     // 0.125 km
-    (1, 250),     // 0.25 km
-    (2, 500),     // 0.5 km
-    (3, 750),     // 0.75 km
-    (4, 1000),    // 1 km
-    (5, 1500),    // 1.5 km
-    (6, 2000),    // 2 km
-    (7, 3000),    // 3 km
-    (8, 4000),    // 4 km
-    (9, 6000),    // 6 km
-    (10, 8000),   // 8 km
-    (11, 12000),  // 12 km
-    (12, 16000),  // 16 km
-    (13, 24000),  // 24 km
-    (14, 32000),  // 32 km
-    (19, 36000),  // 36 km (OUT OF SEQUENCE!)
-    (15, 48000),  // 48 km
-    (20, 64000),  // 64 km (OUT OF SEQUENCE!)
-    (16, 72000),  // 72 km
-    (17, 96000),  // 96 km
-    (18, 120000), // 120 km
-];
-
-/// Furuno wire unit values for the range command protocol
-pub const WIRE_UNIT_NM: i32 = 0;
-pub const WIRE_UNIT_KM: i32 = 1;
-// Wire unit 2 = SM (statute miles), 3 = Kyd (kilo-yards) — not yet implemented
-
-/// Convert meters to Furuno wire index, using the NM table.
-pub fn meters_to_wire_index(meters: i32) -> i32 {
-    lookup_wire_index(&WIRE_INDEX_TABLE, meters)
-}
-
-/// Convert meters to Furuno wire index, using the km table.
-pub fn meters_to_wire_index_km(meters: i32) -> i32 {
-    lookup_wire_index(&WIRE_INDEX_TABLE_KM, meters)
-}
-
-/// Convert meters to Furuno wire index using the appropriate table for the wire unit.
-pub fn meters_to_wire_index_for_unit(meters: i32, wire_unit: i32) -> i32 {
-    match wire_unit {
-        WIRE_UNIT_KM => meters_to_wire_index_km(meters),
-        _ => meters_to_wire_index(meters),
-    }
-}
-
-fn lookup_wire_index(table: &[(i32, i32)], meters: i32) -> i32 {
-    // Return the wire index whose range is closest (in absolute distance) to
-    // the requested value. Previously this was a ceiling search which could
-    // jump to a much larger range (e.g. 66672 m would round up to 72 NM
-    // instead of the nearest 64 NM entry).
-    table
-        .iter()
-        .min_by_key(|(_, m)| (i64::from(*m) - i64::from(meters)).abs())
-        .map(|(idx, _)| *idx)
-        .unwrap_or(15)
-}
-
-/// Convert Furuno wire index to meters (NM mode)
-pub fn wire_index_to_meters(wire_index: i32) -> Option<i32> {
-    WIRE_INDEX_TABLE
-        .iter()
-        .find(|(idx, _)| *idx == wire_index)
-        .map(|(_, meters)| *meters)
-}
-
-/// Convert Furuno wire index to meters (km mode)
-pub fn wire_index_to_meters_km(wire_index: i32) -> Option<i32> {
-    WIRE_INDEX_TABLE_KM
-        .iter()
-        .find(|(idx, _)| *idx == wire_index)
-        .map(|(_, meters)| *meters)
-}
-
-/// Convert Furuno wire index to meters using the appropriate table for the wire unit.
-pub fn wire_index_to_meters_for_unit(wire_index: i32, wire_unit: i32) -> Option<i32> {
-    match wire_unit {
-        WIRE_UNIT_KM => wire_index_to_meters_km(wire_index),
-        _ => wire_index_to_meters(wire_index),
-    }
-}
-
-/// Determine the wire unit for a range value in meters.
-/// Metric distances (km-based) use wire unit 1, nautical use wire unit 0.
-pub fn wire_unit_for_meters(meters: i32) -> i32 {
-    // Check if this is a km-table value by looking for exact match
-    if WIRE_INDEX_TABLE_KM.iter().any(|(_, m)| *m == meters) {
-        // Could be either — check if it's metric (km-based round numbers)
-        if crate::radar::range::Range::is_metric_distance(meters) {
-            return WIRE_UNIT_KM;
-        }
-    }
-    WIRE_UNIT_NM
-}
 
 pub(crate) struct Command {
     key: String,
