+++ src/lib/brand/furuno/mod.rs
@@ -1,11 +1,8 @@
 use bincode::deserialize;
 use log::log_enabled;
-use serde::Deserialize;
 use std::collections::HashMap;
-use std::fmt::{self, Display};
 use std::io::{self, Read, Write};
-use std::net::{IpAddr, Ipv4Addr, SocketAddr, SocketAddrV4};
-use std::time::Duration;
+use std::net::{Ipv4Addr, SocketAddrV4};
 use tokio_graceful_shutdown::{SubsystemBuilder, SubsystemHandle};
 
 use crate::locator::LocatorAddress;
@@ -16,176 +13,21 @@ use crate::{Brand, Cli};
 use super::{LocatorId, RadarLocator};
 
 mod command;
+mod protocol;
 mod report;
 mod settings;
 
-const FURUNO_SPOKES: usize = 8192;
-
-// Maximum supported length of a spoke in pixels. Must be at least as large
-// as any radar's native sweep_len; the DRS4D-NXT reports sweep_len=884, and
-// sizing FURUNO_SPOKE_LEN below the actual spoke length causes each spoke to
-// overflow into the next angle's slot in the GUI's per-angle render buffer,
-// producing a slowly rotating ring/moiré interference pattern across the PPI.
-// 1024 is a round upper bound that accommodates known Furuno models and
-// leaves headroom for future ones; shorter-native spokes are stretched to
-// this size by stretch_spoke so the inner/outer scale stays consistent.
-const FURUNO_SPOKE_LEN: usize = 1024;
-
-const FURUNO_BASE_PORT: u16 = 10000;
-const FURUNO_BEACON_PORT: u16 = FURUNO_BASE_PORT + 10;
-const FURUNO_DATA_PORT: u16 = FURUNO_BASE_PORT + 24;
-
-const FURUNO_BEACON_ADDRESS: SocketAddr = SocketAddr::new(
-    IpAddr::V4(Ipv4Addr::new(172, 31, 255, 255)),
-    FURUNO_BEACON_PORT,
-);
-const FURUNO_DATA_BROADCAST_ADDRESS: SocketAddrV4 =
-    SocketAddrV4::new(Ipv4Addr::new(172, 31, 255, 255), FURUNO_DATA_PORT);
-
-const FURUNO_ANNOUNCE_MAYARA_PACKET: [u8; 32] = [
-    0x1, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1, 0x0, 0x18, 0x1, 0x0, 0x0, 0x0, b'M', b'A',
-    b'Y', b'A', b'R', b'A', 0x0, 0x0, 0x1, 0x1, 0x0, 0x2, 0x0, 0x1, 0x0, 0x12,
-];
-const FURUNO_REQUEST_BEACON_PACKET: [u8; 16] = [
-    0x01, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x01, 0x01, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00,
-];
-const FURUNO_REQUEST_MODEL_PACKET: [u8; 16] = [
-    0x01, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x14, 0x01, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00,
-];
-
-pub(crate) enum CommandMode {
-    Set,
-    Request,
-    New,
-    X,
-    E,
-    O,
-}
-
-impl CommandMode {
-    fn to_char(&self) -> char {
-        match self {
-            CommandMode::Set => 'S',
-            CommandMode::Request => 'R',
-            CommandMode::New => 'N',
-            CommandMode::X => 'X',
-            CommandMode::E => 'E',
-            CommandMode::O => 'O',
-            // Add more cases as needed
-        }
-    }
-}
-
-impl From<u8> for CommandMode {
-    fn from(item: u8) -> Self {
-        match item {
-            b'S' => CommandMode::Set,
-            b'R' => CommandMode::Request,
-            b'N' => CommandMode::New,
-            b'X' => CommandMode::X,
-            b'E' => CommandMode::E,
-            b'O' => CommandMode::O,
-            _ => CommandMode::New,
-        }
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum RadarModel {
-    Unknown,
-    FAR21x7,
-    DRS,
-    FAR14x7,
-    DRS4DL,
-    FAR3000,
-    DRS4DNXT,
-    DRS6ANXT,
-    DRS6AXCLASS,
-    FAR15x3,
-    FAR14x6,
-    DRS4W,
-    DRS12ANXT,
-    DRS25ANXT,
-}
-
-impl Display for RadarModel {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let s: &'static str = match self {
-            RadarModel::Unknown => "Unknown",
-            RadarModel::FAR21x7 => "FAR21x7",
-            RadarModel::DRS => "DRS",
-            RadarModel::FAR14x7 => "FAR14x7",
-            RadarModel::DRS4DL => "DRS4DL",
-            RadarModel::FAR3000 => "FAR3000",
-            RadarModel::DRS4DNXT => "DRS4DNXT",
-            RadarModel::DRS6ANXT => "DRS6ANXT",
-            RadarModel::DRS6AXCLASS => "DRS6AXCLASS",
-            RadarModel::FAR15x3 => "FAR15x3",
-            RadarModel::FAR14x6 => "FAR14x6",
-            RadarModel::DRS4W => "DRS4W",
-            RadarModel::DRS12ANXT => "DRS12ANXT",
-            RadarModel::DRS25ANXT => "DRS25ANXT",
-        };
-
-        write!(f, "{}", s)
-    }
-}
-
-// DRS-4D NXT
-// [01, 00, 00, 01, 00, 00, 00, 00, 00, 01, 00, 18, 01, 00, 00, 00, 52, 44, 30, 30, 33, 32, 31, 32, 01, 01, 00, 02, 00, 01, 00, 12] len 32
-// [ .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   R   D   0   0   3   2   1   2   .   .   .   .   .   .   .   .]
-//                                               ^__length           ^_name, always 8 long?
-// FAR 2127
-// [01, 00, 00, 01, 00, 00, 00, 00, 00, 01, 00, 1A, 01, 00, 00, 00, 52, 41, 44, 41, 52, 00, 00, 00, 01, 00, 00, 03, 00, 01, 00, 04, 00, 05] len 34
-// [ .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   R   A   D   A   R   .   .   .   .   .   .   .   .   .   .   .   .   .]
-//
-// TimeZero
-// [01, 00, 00, 01, 00, 00, 00, 00, 00, 01, 00, 1C, 01, 00, 00, 00, 4D, 46, 30, 30, 33, 31, 35, 30, 01, 01, 00, 04, 00, 0B, 00, 15, 00, 14, 00, 16] len 36
-// [ .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   .   M   F   0   0   3   1   5   0   .   .   .   .   .   .   .   .   .   .   .   .]
-#[derive(Deserialize, Debug, Copy, Clone)]
-#[repr(packed)]
-struct FurunoRadarReport {
-    _header: [u8; 11],
-    length: u8,
-    _filler2: [u8; 4],
-    name: [u8; 8],
-    // ... followed by unknown stuff, looks like two bytes a piece. No idea what they are...
-}
-
-const FURUNO_RADAR_REPORT_HEADER: [u8; 11] =
-    [0x1, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1, 0x0];
-const FURUNO_RADAR_REPORT_LENGTH_MIN: usize = std::mem::size_of::<FurunoRadarReport>();
-
-#[derive(Deserialize, Debug, Copy, Clone)]
-#[repr(packed)]
-// Length 170 bytes
-struct FurunoRadarModelReport {
-    _filler1: [u8; 24],
-    model: [u8; 32],
-    _firmware_versions: [u8; 32],
-    _firmware_version: [u8; 32],
-    serial_no: [u8; 32],
-    _filler2: [u8; 18],
-}
-
-const LOGIN_TIMEOUT: Duration = Duration::from_millis(500);
+use protocol::{
+    ANNOUNCE_MAYARA_PACKET, BASE_PORT, BEACON_ADDRESS, BEACON_REPORT_HEADER,
+    BEACON_REPORT_LENGTH_MIN, DATA_PORT, FurunoRadarModelReport, FurunoRadarReport,
+    LOGIN_EXPECTED_HEADER, LOGIN_MESSAGE, LOGIN_TIMEOUT, MODEL_REPORT_LENGTH, PIXEL_VALUES,
+    RadarModel, REQUEST_BEACON_PACKET, REQUEST_MODEL_PACKET, SPOKE_DATA_MULTICAST_ADDRESS, SPOKES,
+    SPOKE_LEN,
+};
 
 fn login_to_radar(radar_addr: SocketAddrV4) -> Result<u16, io::Error> {
     let mut stream =
         std::net::TcpStream::connect_timeout(&std::net::SocketAddr::V4(radar_addr), LOGIN_TIMEOUT)?;
-
-    // fnet.dll function "login_via_copyright"
-    // From the 13th byte the message is:
-    // "COPYRIGHT (C) 2001 FURUNO ELECTRIC CO.,LTD. "
-    const LOGIN_MESSAGE: [u8; 56] = [
-        //                                              v- this byte is the only variable one
-        0x8, 0x1, 0x0, 0x38, 0x1, 0x0, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x43, 0x4f, 0x50, 0x59, 0x52,
-        0x49, 0x47, 0x48, 0x54, 0x20, 0x28, 0x43, 0x29, 0x20, 0x32, 0x30, 0x30, 0x31, 0x20, 0x46,
-        0x55, 0x52, 0x55, 0x4e, 0x4f, 0x20, 0x45, 0x4c, 0x45, 0x43, 0x54, 0x52, 0x49, 0x43, 0x20,
-        0x43, 0x4f, 0x2e, 0x2c, 0x4c, 0x54, 0x44, 0x2e, 0x20,
-    ];
-    const EXPECTED_HEADER: [u8; 8] = [0x9, 0x1, 0x0, 0xc, 0x1, 0x0, 0x0, 0x0];
 
     stream.set_write_timeout(Some(LOGIN_TIMEOUT))?;
     stream.set_read_timeout(Some(LOGIN_TIMEOUT))?;
@@ -195,7 +37,7 @@ fn login_to_radar(radar_addr: SocketAddrV4) -> Result<u16, io::Error> {
     let mut buf: [u8; 8] = [0; 8];
     stream.read_exact(&mut buf)?;
 
-    if buf != EXPECTED_HEADER {
+    if buf != LOGIN_EXPECTED_HEADER {
         return Err(io::Error::new(
             io::ErrorKind::Other,
             format!("Unexpected reply {:?}", buf),
@@ -203,7 +45,7 @@ fn login_to_radar(radar_addr: SocketAddrV4) -> Result<u16, io::Error> {
     }
     stream.read_exact(&mut buf[0..4])?;
 
-    let port = FURUNO_BASE_PORT + ((buf[0] as u16) << 8) + buf[1] as u16;
+    let port = BASE_PORT + ((buf[0] as u16) << 8) + buf[1] as u16;
     log::debug!(
         "Furuno radar logged in; using port {} for report/command data",
         port
@@ -262,7 +104,7 @@ impl FurunoLocator {
                     Ok(p) => p,
                 }
             } else {
-                FURUNO_DATA_PORT
+                DATA_PORT
             };
             if port != info.send_command_addr.port() {
                 // Furuno radars use a single TCP/IP connection to send commands and
@@ -335,12 +177,12 @@ impl FurunoLocator {
             log::debug!("{}: printable:     {}", from, PrintableSlice::new(report));
         }
 
-        if report.len() >= FURUNO_RADAR_REPORT_LENGTH_MIN
+        if report.len() >= BEACON_REPORT_LENGTH_MIN
             && report[16] == b'R'
-            && report[0..11] == FURUNO_RADAR_REPORT_HEADER
+            && report[0..11] == BEACON_REPORT_HEADER
         {
             self.process_beacon_report(report, from, via)
-        } else if report.len() == 170 {
+        } else if report.len() == MODEL_REPORT_LENGTH {
             self.process_beacon_model_report(report, from, via, radars, subsys)
         } else {
             Ok(())
@@ -419,9 +261,7 @@ impl FurunoLocator {
                     return Ok(());
                 }
 
-                // DRS: spoke data all on a well-known address
-                let spoke_data_addr: SocketAddrV4 =
-                    SocketAddrV4::new(Ipv4Addr::new(239, 255, 0, 2), FURUNO_DATA_PORT);
+                let spoke_data_addr = SPOKE_DATA_MULTICAST_ADDRESS;
 
                 let report_addr: SocketAddrV4 = SocketAddrV4::new(*from.ip(), 0); // Port is set in login_to_radar
                 let send_command_addr: SocketAddrV4 = report_addr.clone();
@@ -437,9 +277,9 @@ impl FurunoLocator {
                     Brand::Furuno,
                     serial_no,
                     dual_suffix,
-                    64,
-                    FURUNO_SPOKES,
-                    FURUNO_SPOKE_LEN,
+                    PIXEL_VALUES,
+                    SPOKES,
+                    SPOKE_LEN,
                     *from,
                     nic_addr.clone(),
                     spoke_data_addr,
@@ -467,9 +307,9 @@ impl FurunoLocator {
                         Brand::Furuno,
                         serial_no,
                         Some("B"),
-                        64,
-                        FURUNO_SPOKES,
-                        FURUNO_SPOKE_LEN,
+                        PIXEL_VALUES,
+                        SPOKES,
+                        SPOKE_LEN,
                         *from,
                         nic_addr.clone(),
                         spoke_data_addr,
@@ -511,12 +351,12 @@ pub(super) fn new(args: &Cli, addresses: &mut Vec<LocatorAddress>) {
     if !addresses.iter().any(|i| i.id == LocatorId::Furuno) {
         addresses.push(LocatorAddress::new(
             LocatorId::Furuno,
-            &FURUNO_BEACON_ADDRESS,
+            &BEACON_ADDRESS,
             Brand::Furuno,
             vec![
-                &FURUNO_REQUEST_BEACON_PACKET,
-                &FURUNO_REQUEST_MODEL_PACKET,
-                &FURUNO_ANNOUNCE_MAYARA_PACKET,
+                &REQUEST_BEACON_PACKET,
+                &REQUEST_MODEL_PACKET,
+                &ANNOUNCE_MAYARA_PACKET,
             ],
             Box::new(FurunoLocator::new(args.clone())),
         ));
